% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rayshader_coaster.R
\name{generate_rayshader_coaster}
\alias{generate_rayshader_coaster}
\title{Generate Rayshader Rollercoaster}
\usage{
generate_rayshader_coaster(
  frames = 360,
  closed = TRUE,
  viewer_offset = 5,
  track_radius = 1,
  track_material = rayrender::diffuse(color = "red"),
  posts = TRUE,
  post_interval = 10,
  post_material = rayrender::diffuse(color = "grey10"),
  light = TRUE,
  lightdirection = 315,
  lightaltitude = 45,
  lightsize = NULL,
  lightintensity = 500,
  lightcolor = "white",
  obj_material = rayrender::diffuse(),
  cache_filename = NULL,
  width = NULL,
  height = NULL,
  ground_material = rayrender::diffuse(),
  ground_size = 10000,
  scene_elements = NULL,
  ...
)
}
\arguments{
\item{frames}{Default `360`. Number of frames in the animation.}

\item{closed}{Default `TRUE`. Whether to close the loop in the rollercoaster.}

\item{viewer_offset}{Default `5`. Height of the viewer along the track.}

\item{track_radius}{Default `1`. Radius of the track.}

\item{track_material}{Default `diffuse(color="red")`. Material for the track.}

\item{posts}{Default `TRUE`. Whether to include posts holding the track up.}

\item{post_interval}{Default `10`. Keyframe interval between posts supporting track.}

\item{post_material}{Default `diffuse(color="grey10")`. Material for the track.}

\item{light}{Default `TRUE`. Whether there should be a light in the scene. If not, the scene will be lit with a bluish sky.}

\item{lightdirection}{Default `315`. Position of the light angle around the scene.
If this is a vector longer than one, multiple lights will be generated (using values from
`lightaltitude`, `lightintensity`, and `lightcolor`)}

\item{lightaltitude}{Default `45`. Angle above the horizon that the light is located.
If this is a vector longer than one, multiple lights will be generated (using values from
`lightdirection`, `lightintensity`, and `lightcolor`)}

\item{lightsize}{Default `NULL`. Radius of the light(s). Automatically chosen, but can be set here by the user.}

\item{lightintensity}{Default `500`. Intensity of the light.}

\item{lightcolor}{Default `white`. The color of the light.}

\item{obj_material}{Default `rayrender::diffuse()`. The material properties of the object file.}

\item{cache_filename}{Name of temporary filename to store OBJ file, if the user does not want to rewrite the file each time.}

\item{width}{Defaults to the width of the rgl window. Width of the rendering.}

\item{height}{Defaults to the height of the rgl window. Height of the rendering.}

\item{ground_material}{Default `diffuse()`. Material defined by the rayrender material functions.}

\item{ground_size}{Default `100000`. The width of the plane representing the ground.}

\item{scene_elements}{Default `NULL`. Extra scene elements to add to the scene, created with rayrender.}

\item{...}{Additional parameters to pass to `rayrender::render_scene()`}
}
\value{
A list with the resulting rayrender scene and the motion keyframes to pass to `rayrender::render_animation()`
}
\description{
Generate Rayshader Rollercoaster
}
