% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rayrender_scene.R
\name{get_rayrender_motion}
\alias{get_rayrender_motion}
\title{Get Animation Info}
\usage{
get_rayrender_motion()
}
\value{
Data frame
}
\description{
Extract the motion information for the rollercoaster scene.
}
\examples{
#This will return an empty data frame if no rollercoaster has been created.
get_rayrender_motion()
}
