% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rayrender_scene.R
\name{get_rayrender_scene}
\alias{get_rayrender_scene}
\title{Get Scene Info}
\usage{
get_rayrender_scene()
}
\value{
Data frame
}
\description{
Extract the rayrender scene with the rollercoaster (to render on it's own).
}
\examples{
#This will return an empty data frame if no rollercoaster has been created.
get_rayrender_scene()
}
