% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rayrender_coaster.R
\name{generate_rayrender_coaster}
\alias{generate_rayrender_coaster}
\title{Generate Rayrender Rollercoaster}
\usage{
generate_rayrender_coaster(
  scene,
  frames = 360,
  closed = TRUE,
  viewer_offset = 5,
  track_radius = 1,
  track_material = rayrender::diffuse(color = "red"),
  posts = TRUE,
  post_interval = 10,
  post_material = rayrender::diffuse(color = "grey10"),
  ...
)
}
\arguments{
\item{scene}{A rayrender scene.}

\item{frames}{Default `360`. Number of frames in the animation.}

\item{closed}{Default `TRUE`. Whether to close the loop in the rollercoaster.}

\item{viewer_offset}{Default `5`. Height of the viewer along the track.}

\item{track_radius}{Default `1`. Radius of the track.}

\item{track_material}{Default `diffuse(color="red")`. Material for the track.}

\item{posts}{Default `TRUE`. Whether to include posts holding the track up.}

\item{post_interval}{Default `10`. Keyframe interval between posts supporting track.}

\item{post_material}{Default `diffuse(color="grey10")`. Material for the track.}

\item{...}{Additional parameters to pass to `rayrender::render_scene()`.}
}
\value{
A list with the resulting rayrender scene and the motion keyframes to pass to `rayrender::render_animation()`
}
\description{
Generate Rayrender Rollercoaster
}
