% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rayshader_coaster.R
\name{generate_rayshader_coaster}
\alias{generate_rayshader_coaster}
\title{Generate Rayshader Rollercoaster}
\usage{
generate_rayshader_coaster(
  frames = 360,
  closed = TRUE,
  viewer_offset = NA,
  track_radius = NA,
  track_material = rayrender::diffuse(color = "red"),
  posts = TRUE,
  post_interval = 10,
  post_material = rayrender::diffuse(color = "grey10"),
  light = TRUE,
  lightdirection = 315,
  lightaltitude = 45,
  lightsize = NULL,
  lightintensity = 500,
  lightcolor = "white",
  obj_material = rayrender::diffuse(),
  cache_filename = NULL,
  width = NULL,
  height = NULL,
  ground_material = rayrender::diffuse(),
  ground_size = 10000,
  scene_elements = NULL,
  ...
)
}
\arguments{
\item{frames}{Default `360`. Number of frames in the animation.}

\item{closed}{Default `TRUE`. Whether to close the loop in the rollercoaster.}

\item{viewer_offset}{Default `NA`. Height of the viewer along the track.  Defaults to twice the `track_radius`.}

\item{track_radius}{Default `NA`. Radius of the track. Defaults to 1/100th the minimum bounding box of the motion.}

\item{track_material}{Default `diffuse(color="red")`. Material for the track.}

\item{posts}{Default `TRUE`. Whether to include posts holding the track up.}

\item{post_interval}{Default `10`. Keyframe interval between posts supporting track.}

\item{post_material}{Default `diffuse(color="grey10")`. Material for the track.}

\item{light}{Default `TRUE`. Whether there should be a light in the scene. If not, the scene will be lit with a bluish sky.}

\item{lightdirection}{Default `315`. Position of the light angle around the scene.
If this is a vector longer than one, multiple lights will be generated (using values from
`lightaltitude`, `lightintensity`, and `lightcolor`)}

\item{lightaltitude}{Default `45`. Angle above the horizon that the light is located.
If this is a vector longer than one, multiple lights will be generated (using values from
`lightdirection`, `lightintensity`, and `lightcolor`)}

\item{lightsize}{Default `NULL`. Radius of the light(s). Automatically chosen, but can be set here by the user.}

\item{lightintensity}{Default `500`. Intensity of the light.}

\item{lightcolor}{Default `white`. The color of the light.}

\item{obj_material}{Default `rayrender::diffuse()`. The material properties of the object file.}

\item{cache_filename}{Name of temporary filename to store OBJ file, if the user does not want to rewrite the file each time.}

\item{width}{Defaults to the width of the rgl window. Width of the rendering.}

\item{height}{Defaults to the height of the rgl window. Height of the rendering.}

\item{ground_material}{Default `diffuse()`. Material defined by the rayrender material functions.}

\item{ground_size}{Default `100000`. The width of the plane representing the ground.}

\item{scene_elements}{Default `NULL`. Extra scene elements to add to the scene, created with rayrender.}

\item{...}{Additional parameters to pass to `rayrender::render_scene()`}
}
\value{
A list with the resulting rayrender scene and the motion keyframes to pass to `rayrender::render_animation()`
}
\description{
Generate Rayshader Rollercoaster
}
\examples{
#Generate a rayshader scene
\donttest{
library(rayshader)

volcano |>
  sphere_shade() |>
  plot_3d(volcano,fov=90)

#Now, let's fly through a series of points to build the rollercoaster.
#You can pass arguments to `render_scene()` via `...`.
#This function doesn't return anything, but saves the scene with the rollercoaster and the
#motion information internally. This extracts the scene from `rayshader` directly from `rgl`.

#Here, we turn off the lighting and use ambient lighting only
generate_rayshader_coaster(light=FALSE, ambient_light=T)

#We can also add extra lights by passing arguments via `...` to `rayshader::render_highquality()`
generate_rayshader_coaster(lightdirection=c(315,45),lightcolor=c("orange","white"))

#We now can create our animation using `animate_rollercoaster()`. This extracts the information
#(scene and motion) in the previous step. If no filename is given,
#it will simply preview the animation in the interactive screen, without saving anything.
#Set samples to `1` to render a quick preview.
animate_rollercoaster(samples=1, width=200,height=200)

#Increase the number of samples and resolution for a high quality animation (but longer render):
animate_rollercoaster(samples=128, width=800, height=800)

#Alternatively, get the scene and motion data yourself and call `rayrender::render_animation()`
#directly.
scene = get_rayrender_scene()
motion = get_rayrender_motion()

render_animation(scene, motion, samples=1, sample_method="sobol_blue")
}
}
